Moved _validate_tile_shape to separate file
